test_python/requirements.txt
numpy>=1.21
Pillow>=9.0
opencv-python>=4.5
